requirements.txt
Babel==2.6.0
Django==2.1.2
django-environ==0.4.5
django-extensions==2.1.3
django-phonenumber-field==2.0.1
djangorestframework==3.9.0
phonenumbers==8.9.16
pip==18.0
pytz==2018.5
setuptools==40.4.3
six==1.11.0
wheel==0.32.0
mysqlclient